Reinstate uos_depth_est_old.py
